+++ plugins/modules/zhmc_partition.py
@@ -494,7 +494,8 @@ from operator import itemgetter  # noqa: E402
 from ..module_utils.common import log_init, Error, ParameterError, \
     StatusError, stop_partition, start_partition, \
     wait_for_transition_completion, eq_hex, get_hmc_auth, get_session, \
-    to_unicode, process_normal_property, missing_required_lib  # noqa: E402
+    to_unicode, process_normal_property, missing_required_lib, \
+    common_fail_on_import_errors  # noqa: E402
 
 try:
     import requests.packages.urllib3
@@ -1912,6 +1913,8 @@ def main():
     if not IMP_ZHMCCLIENT:
         module.fail_json(msg=missing_required_lib("zhmcclient"),
                          exception=IMP_ZHMCCLIENT_ERR)
+
+    common_fail_on_import_errors(module)
 
     log_file = module.params['log_file']
     log_init(LOGGER_NAME, log_file)